神魔竞技-下午挑战高战力

+++ project/Sxd.cs
@@ -411,7 +411,9 @@ namespace 神仙道
                                 var _challengeList = ((JArray)response[0]).ToList();
                                 // is_challenge, NO:int = 27;  .Where(x=>(byte)x[6]==27)
                                 _challengeList.Sort((x, y) => ((int)x[5]).CompareTo((int)y[5]));
-                                _challengeList.RemoveRange(2, 3);
+                                // 如果是上午，则去掉后3个战力强的
+                                if (DateTime.Now.Hour < 12)
+                                    _challengeList.RemoveRange(2, 3);
                                 _challengeList = _challengeList.Where(x => (byte)x[6] == 27).ToList();
                                 if (_challengeList.Count <= 0)
                                     break;
@@ -451,7 +453,7 @@ namespace 神仙道
                             // SUCCESS:int = 28;
                             if ((byte)response[0] == 28)
                                 Logger.Log(string.Format("【神魔竞技】领取排名奖励铜钱[{0}]，声望[{1}]，血脉精华[{2}]", response[1], response[2], response[3]));
-                        }
+                        }// 领取排名奖励
 
                         // 神魔大礼
                         // public static const StSuperSportScoreAward:Array = [[1, "普通宝箱", 1100, 5, 20, 300000, 500, 2], [2, "普通宝箱", 1300, 8, 50, 500000, 600, 2], [3, "普通宝箱", 1500, 10, 100, 1000000, 800, 2], [4, "普通宝箱", 1800, 20, 200, 1200000, 1000, 2], [5, "普通宝箱", 2000, 30, 300, 1500000, 1200, 2], [6, "青铜宝箱", 2200, 50, 500, 2000000, 1500, 3], [7, "白银宝箱", 2400, 80, 700, 3000000, 1600, 4], [8, "黄金宝箱", 2600, 100, 1000, 5000000, 1800, 5]];
@@ -487,7 +489,7 @@ namespace 神仙道
                                 else
                                     Logger.Log("【神魔竞技】[神魔大礼]领取错误", ConsoleColor.Red);
                             }
-                        }
+                        }// 神魔大礼
 
                     } //if (functionIds.Contains(93)) // 93:["st_super_sport","206","仙界竞技场"],，实际上是神魔竞技
                 }//if ((int)response[0] == 0)
